Extract shared yt-dlp settings into SHARED_YT_DLP_SETTINGS
