+++ downloader/twitch_downloader.py
@@ -8,6 +8,7 @@ from injector import inject
 from config import Config
 from downloader import Downloader
 from yt_dlp_logger import YtDlpLogger
+from yt_dlp_settings import SHARED_YT_DLP_SETTINGS
 
 
 class TwitchDownloader(Downloader):
@@ -20,11 +21,9 @@ class TwitchDownloader(Downloader):
         if not url:
             raise ValueError("url is required")
 
-        def _retry_sleep(n):  # noqa: ARG001
-            return 15
-
         def sync():
             ydl_opts = {
+                **SHARED_YT_DLP_SETTINGS,
                 "logger": self._yt_dlp_logger,
                 "format": "best",
                 "merge_output_format": "mp4",
@@ -33,24 +32,6 @@ class TwitchDownloader(Downloader):
                     self._config["output_folder"],
                     "[%(uploader)s] %(title)s.%(ext)s",
                 ),
-                # --- The Retry Trio ---
-                "retries": 10,  # Generic network retries
-                "fragment_retries": 10,  # Video chunk/fragment retries
-                "extractor_retries": 10,  # Website parsing/scraping retries
-                "file_access_retries": 10,  # Local disk/NAS access retries
-                # Sleep between each retry attempt (all retry types)
-                "retry_sleep_functions": {
-                    "http": _retry_sleep,
-                    "fragment": _retry_sleep,
-                    "file_access": _retry_sleep,
-                    "extractor": _retry_sleep,
-                },
-                # --- Precise Timing Control ---
-                "sleep_interval": 15,  # Seconds to wait between download tasks
-                "max_sleep_interval": 15,  # Keep it strictly at 15s (no randomization)
-                "sleep_requests": 5,  # Wait 5s between finding info for each video
-                # --- Safety Buffers ---
-                "socket_timeout": 30,  # Wait 30s before considering a socket "dead"
                 # Suppress FFmpeg's direct stderr output (it bypasses yt-dlp's logger)
                 "postprocessor_args": {"ffmpeg": ["-loglevel", "warning", "-stats", "-stats_period", "60"]},
                 "external_downloader_args": {"ffmpeg": ["-loglevel", "warning", "-stats", "-stats_period", "60"]},
